Create .bash_profile
If you have used UNIX for a while, you will know that there are many commands available and that some of them have very cryptic names and/or can be invoked with a truckload of options and arguments. So, it would be nice to have a feature allowing you to rename these commands or type something simple instead of a list of options. Bash provides such a feature : the alias .
Aliasses can be defined on the command line, in .bash_profile

The .bash_profile you placed in your home directory and every time that you add or change your aliases you simple run in the Terminal:
 . ~/.bash_profile 
That will refresh all the changes and make new alias to be available.